+++ stack-graphs/tests/it/c/mod.rs
@@ -8,6 +8,7 @@
 mod can_create_graph;
 mod can_find_local_nodes;
 mod can_find_partial_paths_in_file;
+mod can_find_qualified_definitions_with_phased_partial_path_stitching;
 mod can_jump_to_definition;
 mod can_jump_to_definition_with_phased_partial_path_stitching;
 mod can_jump_to_definition_with_phased_path_stitching;